Rename cypress-gatsby into gatsby-cypress
