+++ e2e-tests/development-runtime/cypress/support/index.js
@@ -1,2 +1,2 @@
-import "../../../../packages/cypress-gatsby"
+import "../../../../packages/gatsby-cypress"
 import "./commands"
