+++ e2e-tests/gatsby-image/cypress/support/index.js
@@ -14,8 +14,8 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
-// TODO: import as "cypress-gatsby" once this is published to NPM
-import '../../../../packages/cypress-gatsby'
+// TODO: import as "gatsby-cypress" once this is published to NPM
+import '../../../../packages/gatsby-cypress'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
