+++ packages/gatsby-cypress/CHANGELOG.md
@@ -1,0 +1,36 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="0.1.3"></a>
+
+## [0.1.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands/compare/gatsby-cypress@0.1.2...gatsby-cypress@0.1.3) (2018-12-24)
+
+### Bug Fixes
+
+- **gatsby:** handle missing pages metadata chunk ([#10507](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands/issues/10507)) ([b9411d8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands/commit/b9411d8))
+
+<a name="0.1.2"></a>
+
+## [0.1.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands/compare/gatsby-cypress@0.1.1...gatsby-cypress@0.1.2) (2018-10-02)
+
+**Note:** Version bump only for package gatsby-cypress
+
+<a name="0.1.1"></a>
+
+## [0.1.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands/compare/gatsby-cypress@0.1.1-rc.1...gatsby-cypress@0.1.1) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-cypress
+
+<a name="0.1.1-rc.1"></a>
+
+## [0.1.1-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands/compare/gatsby-cypress@0.1.1-rc.0...gatsby-cypress@0.1.1-rc.1) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-cypress
+
+<a name="0.1.1-rc.0"></a>
+
+## 0.1.1-rc.0 (2018-09-11)
+
+**Note:** Version bump only for package gatsby-cypress
