+++ packages/gatsby-cypress/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "cypress-gatsby",
+  "name": "gatsby-cypress",
   "version": "0.1.3",
   "description": "Cypress tools for Gatsby projects",
   "main": "index.js",
-  "repository": "https://github.com/gatsbyjs/gatsby/tree/master/packages/cypress-gatsby-commands#readme",
+  "repository": "https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cypress-commands#readme",
   "author": "David Bailey <david.j.b@vivaldi.net>",
   "license": "MIT",
   "bugs": {
