+++ packages/gatsby/cache-dir/api-runner-browser.js
@@ -6,7 +6,7 @@ const {
 } = require(`./loader`).publicLoader
 
 exports.apiRunner = (api, args = {}, defaultReturn, argTransform) => {
-  // Hooks for cypress-gatsby's API handler
+  // Hooks for gatsby-cypress's API handler
   if (process.env.CYPRESS_SUPPORT) {
     if (window.___apiHandler) {
       window.___apiHandler(api)
